Hide autocompleter suggestions on message submit
This adds a call to `MessageComposer.hideSuggestions()` in the
`onMessageSubmit` function so that the autocompleter is always closed after a
message is sent.

To Reproduce
------------------

On master, create a message like ":D" and send the message. This will trigger
the autocompleter to show up and suggest emojis, but after the message is sent,
the autocompleter will still be visible.

On this branch, sending the same message will close the autocompleter after the
message is sent.

Fix #1203