allow editting mentions text for users with usernames
when now holding the Alt key while selecting a user from the mentions
autocomplete suggestions the text inserted into the composer will be
forced to be in the

	@1234567 (firstname)

format (i.e. referring to the user by id and local firstname) instead of

	@username

even if the selected user has a username set.
This allows to manually edit the display text ("firstname") of the
mention.